+++ common/commands/src/validator/mixnet/operators/mixnode/rewards/claim_operator_reward.rs
@@ -1,0 +1,20 @@
+// Copyright 2021 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: Apache-2.0
+
+use crate::context::SigningClient;
+use clap::Parser;
+use log::info;
+
+#[derive(Debug, Parser)]
+pub struct Args {}
+
+pub async fn claim_operator_reward(_args: Args, client: SigningClient) {
+    info!("Claim operator reward");
+
+    let res = client
+        .execute_claim_operator_reward(None)
+        .await
+        .expect("failed to claim operator reward");
+
+    info!("Claiming operator reward: {:?}", res)
+}