+++ contracts/vesting/src/contract.rs
@@ -1,8 +1,8 @@
 use crate::errors::ContractError;
-use crate::queued_migrations::migrate_config_from_env;
 use crate::storage::{
-    account_from_address, locked_pledge_cap, update_locked_pledge_cap, BlockTimestampSecs, NodeId,
-    ADMIN, DELEGATIONS, MIXNET_CONTRACT_ADDRESS, MIX_DENOM, OLD_DELEGATIONS,
+    account_from_address, locked_pledge_cap, remove_delegation, save_delegation,
+    update_locked_pledge_cap, BlockTimestampSecs, NodeId, ADMIN, DELEGATIONS,
+    MIXNET_CONTRACT_ADDRESS, MIX_DENOM, OLD_DELEGATIONS,
 };
 use crate::traits::{
     DelegatingAccount, GatewayBondingAccount, MixnodeBondingAccount, VestingAccount,
@@ -46,7 +46,6 @@ pub fn instantiate(
 
 #[entry_point]
 pub fn migrate(_deps: DepsMut<'_>, _env: Env, _msg: MigrateMsg) -> Result<Response, ContractError> {
-    migrate_config_from_env(_deps, _env, _msg)?;
     Ok(Response::default())
 }